modify the TestHiveConnectorTest
